requirements.txt
fastapi==0.118.1
uvicorn==0.22.0
python-dotenv==1.1.1
anthropic==0.69.0
google-genai==1.41.0